speed up labeling with c code

+++ api/api_athena/src/c_api/athn.cpp
@@ -1,14 +1,16 @@
 #include "athn.h"
+#include <omp.h>
 using namespace std;
 athnStatus athn_label_minbars(real64* open, real64* high, real64* low, uint64* secs, int32 num_bars, // input: all min bars
-                              int32* minbar_ids, int32 num_minbar_ids, // input: index to label
+                              uint64* minbar_ids, int32 num_minbar_ids, // input: index to label
                               real64 ret_thd, // input: return threshold
                               int32 ndays, // input: position lifetime
                               int32* labels) {
 // secs is the elapsed time in second relative to the first min bar
 
     uint64 pos_lifetime = ndays*3600*24;
-
+//    omp_set_num_threads(32);
+//#pragma omp parallel for
     for(int32 idx = 0; idx < num_minbar_ids; idx++) {
         int32 tid = minbar_ids[idx];
         real64 p0 = open[tid];
