+++ api/api_athena/src/c_api/athn.h
@@ -12,7 +12,7 @@ typedef unsigned long   uint64;
 athnStatus athn_label_minbars(real64* open, real64* high, real64* low,
                               uint64* secs, //
                               int32 num_bars, // input: number of min bars
-                          int32* minbar_ids, int32 num_minbar_ids, // input: index to label
+                          uint64* minbar_ids, int32 num_minbar_ids, // input: index to label
                           real64 ret_thd, // input: return threshold
                           int32 ndays, // input: position lifetime
                           int32* labels); // output
